Deal with Cloudflare bullshit
Closes #1

This is not a great solution for monitoring purposes though.
Because this results in the service not dying, some other creative way
to track downtime should likely be handled.

I also need to find a way to coordinate this with stackchat.cpp, but I'm
not sure how I want to go about that yet.